About: Add a frag for introducing the crew.
Notes:
* Loading avatar icons from github requires an async task
* Dynamically changing preferences from an asynchronous task has some issues
or limitations in Android. Therefore I had to use static preferences in the
crew.xml, separate the labeling from adding the icon, and further add a delay
in the for-loop to prevent bizarre problems, e.g. sometimes a few items were
not labeled or an icon was missing.

Squashed-in: Crew: Add Our new team member Kyle.

Signed-off-by: spezi77 <spezi7713@gmx.net>
Change-Id: I4c2e80cab72c5097cdeb8cf17139ea12c1053361